Name sqlite DB the same as the Catalog
Previously, the main Catalog resource looked for a sqlite
DB file named after the director with a .db extension.
This caused conflicts when the scripts provided to
manage the sqlite database that are shipped with bacula
created the database since the scripts name the DB file
after the name of the catalog. This commit tells the
catalog to look for a DB file by the same name.