Support creating clients via hash/ENC
Previous to this commit, creating new clients
to be backed up required modifying the
bacula-dir.conf.erb template.

This commit adds a new class parameter,
`clients` that accepts a hash. The key of the
hash is the FQDN of the host while the value
is a hash of parameters. In addition to the
class parameter, users can specify top scope
(ENC) parameters that will generate clients.
The parameter to create in the ENC is of the
format bacula_client_node.domain.com.  The
value of the parameter is in the format
resourceparam=value,resourceparam=value.

So to create a client mybackup.mydomain.com
with to backup with the fileset Basic:noHome
on the Weekly schedule, the name of the ENC
parameter is
`bacula_client_mybackup.mydomain.com` with a
value of
`fileset=Basic:noHome,schedule=Weekly`.
Using the clients class parameter, the value
of $clients would be
$clients = { 'mybackup.mydomain.com' => {
    'fileset'  => 'Basic:noHome',
    'schedule' => 'Weekly',
  }
}